requirements.txt
numpy>=1.16.2
Pillow>=6.0.0
pyphotonfile>=0.2.1
PySide2>=5.12.2
